no more PLUGININFOEX::replacesDefaultModule, that old & nasty clutch since now all Myranda plugins are binary incompatible with Miranda's
git-svn-id: http://svn.miranda-ng.org/main/trunk@743 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/CryptoPP/base16.cpp
@@ -18,7 +18,7 @@ char *base16encode(const char *inBuffer, int count) {
 }
 
 
-char *base16decode(const char *inBuffer, int *count) {
+char *base16decode(const char *inBuffer, size_t *count) {
 
 	char *outBuffer = (char *) malloc(*count);
 	BYTE *outBufferPtr = (BYTE *) outBuffer;
@@ -54,9 +54,9 @@ char *base16decode(const char *inBuffer, int *count) {
 }
 
 
-char *base16decode(const char *inBuffer) {
-
-    int count = (int)strlen(inBuffer);
+char *base16decode(const char *inBuffer)
+{
+	size_t count = strlen(inBuffer);
 	return base16decode(inBuffer, &count);
 }
